Split LuaSandbox into two files
